requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-aiplatform>=1.42.1,<2.0.0
google-cloud-firestore==2.13.1
chromadb==0.4.15
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
sentence-transformers>=2.3.0
huggingface_hub>=0.16.0
transformers>=4.30.0
torch>=2.0.0
tiktoken==0.5.1
PyJWT==2.8.0
cryptography==41.0.7
python-pptx==0.6.21
PyMuPDF==1.23.8
Pillow>=9.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
redis>=4.5.0
httptools==0.6.1
